+++ bevy_tiled/Cargo.toml
@@ -7,6 +7,8 @@ edition = "2021"
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
 [dependencies]
+bevy_asset_loader = { version = "0.12.0", features = ["stageless", "dynamic_assets"] }
+anyhow = "1.0.4"
 thiserror = "1"
 serde = "1"
 bevy = "0.8"
