+++ bevy_tiled/src/lib.rs
@@ -8,4 +8,18 @@ pub mod tiled_map_asset;
 pub mod map_scheme;
 
 pub use tiled_ext::*;
+pub use tiled_map_asset::*;
 //pub use map_scheme::*;
+
+use bevy::prelude::*;
+
+#[derive(Default)]
+pub struct TiledPlugin;
+
+impl Plugin for TiledPlugin {
+    fn build(&self, app: &mut App) {
+        app
+            .add_asset::<TiledMap>()
+            .add_asset_loader(TiledMapLoader);
+    }
+}
